+++ tests/test_investor_metrics_service.py
@@ -1,11 +1,7 @@
 import pytest
 import pandas as pd
-from unittest.mock import AsyncMock
+from unittest.mock import Mock
 from parcllabs import ParclLabsClient
-
-import nest_asyncio
-
-nest_asyncio.apply()
 
 # Mock Data for testing
 mock_response = {"parcl_id": 1, "items": [{"metric": 10}, {"metric": 20}], "links": {}}
@@ -14,26 +10,25 @@ mock_response = {"parcl_id": 1, "items": [{"metric": 10}, {"metric": 20}], "link
 @pytest.fixture
 def client():
     client = ParclLabsClient(api_key="test_api_key")
-    client.investor_metrics.housing_stock_ownership._fetch = AsyncMock(
+    client.investor_metrics.housing_stock_ownership._fetch = Mock(
         return_value=mock_response
     )
-    client.investor_metrics.new_listings_for_sale_rolling_counts._fetch = AsyncMock(
+    client.investor_metrics.new_listings_for_sale_rolling_counts._fetch = Mock(
         return_value=mock_response
     )
-    client.investor_metrics.purchase_to_sale_ratio._fetch = AsyncMock(
+    client.investor_metrics.purchase_to_sale_ratio._fetch = Mock(
         return_value=mock_response
     )
-    client.investor_metrics.housing_event_counts._fetch = AsyncMock(
+    client.investor_metrics.housing_event_counts._fetch = Mock(
         return_value=mock_response
     )
-    client.investor_metrics.housing_event_prices._fetch = AsyncMock(
+    client.investor_metrics.housing_event_prices._fetch = Mock(
         return_value=mock_response
     )
     return client
 
 
-@pytest.mark.asyncio
-async def test_investor_metrics_housing_stock_ownership_retrieve(client):
+def test_investor_metrics_housing_stock_ownership_retrieve(client):
     result = client.investor_metrics.housing_stock_ownership.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -43,8 +38,7 @@ async def test_investor_metrics_housing_stock_ownership_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_investor_metrics_new_listings_for_sale_rolling_counts_retrieve(client):
+def test_investor_metrics_new_listings_for_sale_rolling_counts_retrieve(client):
     result = client.investor_metrics.new_listings_for_sale_rolling_counts.retrieve(
         parcl_ids=[1]
     )
@@ -56,8 +50,7 @@ async def test_investor_metrics_new_listings_for_sale_rolling_counts_retrieve(cl
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_investor_metrics_purchase_to_sale_ratio_retrieve(client):
+def test_investor_metrics_purchase_to_sale_ratio_retrieve(client):
     result = client.investor_metrics.purchase_to_sale_ratio.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -67,8 +60,7 @@ async def test_investor_metrics_purchase_to_sale_ratio_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_investor_metrics_housing_event_counts_retrieve(client):
+def test_investor_metrics_housing_event_counts_retrieve(client):
     result = client.investor_metrics.housing_event_counts.retrieve(parcl_ids=[1])
     assert isinstance(result, pd.DataFrame)
     assert not result.empty
@@ -79,8 +71,7 @@ async def test_investor_metrics_housing_event_counts_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_investor_metrics_housing_event_prices_retrieve(client):
+def test_investor_metrics_housing_event_prices_retrieve(client):
     result = client.investor_metrics.housing_event_prices.retrieve(parcl_ids=[1])
     assert isinstance(result, pd.DataFrame)
     assert not result.empty