+++ tests/test_market_metrics_service.py
@@ -1,10 +1,6 @@
 import pytest
-from unittest.mock import AsyncMock
+from unittest.mock import Mock
 from parcllabs import ParclLabsClient
-
-import nest_asyncio
-
-nest_asyncio.apply()
 
 # Mock Data for testing
 mock_response = {"parcl_id": 1, "items": [{"metric": 10}, {"metric": 20}], "links": {}}
@@ -13,22 +9,17 @@ mock_response = {"parcl_id": 1, "items": [{"metric": 10}, {"metric": 20}], "link
 @pytest.fixture
 def client():
     client = ParclLabsClient(api_key="test_api_key")
-    client.market_metrics.housing_event_prices._fetch = AsyncMock(
-        return_value=mock_response
-    )
-    client.market_metrics.all_cash._fetch = AsyncMock(return_value=mock_response)
-    client.market_metrics.housing_stock._fetch = AsyncMock(return_value=mock_response)
-    client.market_metrics.housing_event_counts._fetch = AsyncMock(
-        return_value=mock_response
-    )
-    client.market_metrics.housing_event_property_attributes._fetch = AsyncMock(
+    client.market_metrics.housing_event_prices._fetch = Mock(return_value=mock_response)
+    client.market_metrics.all_cash._fetch = Mock(return_value=mock_response)
+    client.market_metrics.housing_stock._fetch = Mock(return_value=mock_response)
+    client.market_metrics.housing_event_counts._fetch = Mock(return_value=mock_response)
+    client.market_metrics.housing_event_property_attributes._fetch = Mock(
         return_value=mock_response
     )
     return client
 
 
-@pytest.mark.asyncio
-async def test_market_metrics_housing_event_prices_retrieve(client):
+def test_market_metrics_housing_event_prices_retrieve(client):
     result = client.market_metrics.housing_event_prices.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -38,8 +29,7 @@ async def test_market_metrics_housing_event_prices_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_market_metrics_all_cash_retrieve(client):
+def test_market_metrics_all_cash_retrieve(client):
     result = client.market_metrics.all_cash.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -49,8 +39,7 @@ async def test_market_metrics_all_cash_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_market_metrics_housing_stock_retrieve(client):
+def test_market_metrics_housing_stock_retrieve(client):
     result = client.market_metrics.housing_stock.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -60,8 +49,7 @@ async def test_market_metrics_housing_stock_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_market_metrics_housing_event_counts_retrieve(client):
+def test_market_metrics_housing_event_counts_retrieve(client):
     result = client.market_metrics.housing_event_counts.retrieve(parcl_ids=[1])
     assert not result.empty
     assert "parcl_id" in result.columns
@@ -71,8 +59,7 @@ async def test_market_metrics_housing_event_counts_retrieve(client):
     assert result.iloc[1]["metric"] == 20
 
 
-@pytest.mark.asyncio
-async def test_market_metrics_housing_event_property_attributes_retrieve(client):
+def test_market_metrics_housing_event_property_attributes_retrieve(client):
     result = client.market_metrics.housing_event_property_attributes.retrieve(
         parcl_ids=[1]
     )